Move command entry points to NSwag.Commands
This removes the assembly loader from the reference graph of assemblies
like NSwag.SwaggerGeneration.AspNetCore
